Refactoed the core library
