+++ core/fs.h
@@ -1,11 +1,49 @@
 #ifndef FILESYSTEM_H
 #define FILESYSTEM_H
 
-#ifndef BASE_H
-# include "base.h"
-#endif
+#include "strings.h"
 
-#ifdef BASE_IMPLEMENTATION
+typedef struct {
+  String name;
+  char *extension;
+  i64 size;
+  i64 modifyTime;
+} File;
+
+typedef struct folder_t {
+  String name;
+
+  struct folder_t *folders;
+  size_t folderCount;
+
+  File *files;
+  size_t fileCount;
+
+  size_t totalCount;
+} Folder;
+
+enum FileError { 
+  FILE_NOT_EXIST = 1, 
+  FILE_OPEN_FAILED, 
+  FILE_GET_SIZE_FAILED, 
+  FILE_READ_FAILED,
+  FILE_GET_ATTRIBUTES_FAILED,
+  FILE_WRITE_FAILED,
+  FILE_DELETE_FAILED,
+  FILE_RENAME_FAILED
+};
+
+String GetCwd();
+void SetCwd(String destination);
+Folder *GetDirFiles(String initial);
+Folder *NewFolder();
+void FreeFolder(Folder *folder);
+errno_t FileStats(String *path, File *file);
+errno_t FileRead(String *path, String *result);
+errno_t FileWrite(String *path, String *data);
+errno_t FileDelete(String *path);
+errno_t FileRename(String *oldPath, String *newPath);
+bool Mkdir(String path);
 
 #define MAX_FILES 200
 
@@ -49,5 +87,4 @@ void FreeFolder(Folder *folder) {
 # include "linux/files.h"
 #endif
 
-#endif
 #endif
